♻️ Move `flatten_dict` and `unflatten_dict` to `utils.py`
